Change PersistentVolumeClaimName to a general source for snapshot
